[Setting/#3] 버튼 컴포넌트 생성
