+++ KKAMPPAK/KKAMPPAK/Resources/Assets.xcassets/Onboarding/Gifs/OnboardingBlink.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "OnboardingBlink.gif",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
