+++ KKAMPPAK/KKAMPPAK/Resources/Assets.xcassets/Onboarding/Gifs/OnboardingGif.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "OnboardingGif.gif",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
