+++ KKAMPPAK/KKAMPPAK/Resources/Assets.xcassets/Onboarding/Gifs/OnoboardingSide.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "data" : [
+    {
+      "filename" : "OnoboardingSide.gif",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
