+++ KKAMPPAK/KKAMPPAK/Resources/Assets.xcassets/Onboarding/OnboardingImage1.imageset/Contents.json
@@ -1,0 +1,22 @@
+{
+  "images" : [
+    {
+      "filename" : "img1.png",
+      "idiom" : "universal",
+      "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "scale" : "2x"
+    },
+    {
+      "filename" : "img1 1.png",
+      "idiom" : "universal",
+      "scale" : "3x"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
